refactor: path 명시 컨벤션 수정 및 admin 관련 api 병합
