--- src/main/java/roomescape/page/AdminPageController.java
@@ -1,14 +1,0 @@
-package roomescape.page;
-
-import org.springframework.stereotype.Controller;
-import org.springframework.web.bind.annotation.GetMapping;
-
-@Controller
-public class AdminPageController {
-
-    @GetMapping("/admin")
-    public String admin(
-    ) {
-        return "/admin/index.html";
-    }
-}
