+++ src/test/java/roomescape/admin/AdminControllerTest.java
@@ -1,4 +1,4 @@
-package roomescape.page;
+package roomescape.admin;
 
 import io.restassured.RestAssured;
 import org.junit.jupiter.api.DisplayName;
@@ -8,7 +8,16 @@ import org.springframework.test.annotation.DirtiesContext;
 
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.DEFINED_PORT)
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
-class ReservationPageControllerTest {
+class AdminControllerTest {
+
+    @Test
+    @DisplayName("어드민 페이지 테스트")
+    void adminTest() {
+        RestAssured.given().log().all()
+                .when().get("/admin")
+                .then().log().all()
+                .statusCode(200);
+    }
 
     @Test
     @DisplayName("예약 페이지 테스트")
